+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.4
+- In case of custom templates, do not override the default template layout file
+- Tentative to fix issues [#12](https://github.com/ludorival/vscode-preview-jsdoc/issues/12) [#11](https://github.com/ludorival/vscode-preview-jsdoc/issues/11) [#10](https://github.com/ludorival/vscode-preview-jsdoc/issues/10)
+- exclude node_modules from jsdoc configuration
 ## 1.0.3
 - Fix issue [#6](https://github.com/ludorival/vscode-preview-jsdoc/issues/6)